Merged in floh's code from last month. Notable changes include: - Dynamic memory allocation is now done using various specialized heaps. - Rendering is now done in a separate dedicated thread. - N2 particle system has been removed (but there is no replacement as yet). - Character animation subsystem has undergone some work, but doesn't appear to be complete yet. - Vertex component packing.
A nice overview of the major changes can be found on floh's blog.

Unfortunately the vertex component packing changes are not backwards compatible with existing mesh resources, so meshes need to be re-exported before they can be loaded again. Since we don't have the updated resources as yet, all the Nebula apps that display pretty things on screen currently crash on startup. We hope to rectify the situation in the near future.